requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
python-telegram-bot==20.7
psutil==5.9.6
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
python-telegram-bot[job-queue]